[expo-notifications] Fix types and add web support for NotificationScheduler
