+++ packages/expo-notifications/src/NotificationScheduler.ts
@@ -1,17 +1,5 @@
-import { NativeModulesProxy, ProxyNativeModule } from '@unimodules/core';
+import { NativeModulesProxy } from '@unimodules/core';
 
-import { NotificationTriggerInput } from './NotificationScheduler.types';
-import { NotificationRequest, NotificationContentInput } from './Notifications.types';
-
-export interface NotificationSchedulerModule extends ProxyNativeModule {
-  getAllScheduledNotificationsAsync: () => Promise<NotificationRequest[]>;
-  scheduleNotificationAsync: (
-    identifier: string,
-    notificationContent: NotificationContentInput,
-    trigger: NotificationTriggerInput
-  ) => Promise<string>;
-  cancelScheduledNotificationAsync: (identifier: string) => Promise<void>;
-  cancelAllScheduledNotificationsAsync: () => Promise<void>;
-}
+import { NotificationSchedulerModule } from './NotificationScheduler.types';
 
 export default (NativeModulesProxy.ExpoNotificationScheduler as any) as NotificationSchedulerModule;
