+++ packages/expo-notifications/src/cancelAllScheduledNotificationsAsync.ts
@@ -3,8 +3,8 @@ import { UnavailabilityError } from '@unimodules/core';
 import NotificationScheduler from './NotificationScheduler';
 
 export default async function cancelScheduledNotificationAsync(): Promise<void> {
-  if (!NotificationScheduler.cancelScheduledNotificationAsync) {
-    throw new UnavailabilityError('Notifications', 'cancelScheduledNotificationAsync');
+  if (!NotificationScheduler.cancelAllScheduledNotificationsAsync) {
+    throw new UnavailabilityError('Notifications', 'cancelAllScheduledNotificationsAsync');
   }
 
   return await NotificationScheduler.cancelAllScheduledNotificationsAsync();
